+++ library/feature/issuereporter/src/main/kotlin/com/mihaicristiancondrea/android/libs/apptoolkit/feature/issuereporter/ui/views/IssueDescriptionField.kt
@@ -57,7 +57,6 @@ import com.mihaicristiancondrea.android.libs.apptoolkit.core.common.data.reposit
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.common.domain.models.analytics.AnalyticsValue
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.common.utils.constants.ui.SizeConstants
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.designsystem.ui.icons.ToolkitIcon
-import com.mihaicristiancondrea.android.libs.apptoolkit.core.ui.views.buttons.ButtonMeasurements
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.ui.views.buttons.GeneralButton
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.ui.views.buttons.GeneralButtonStyle
 import com.mihaicristiancondrea.android.libs.apptoolkit.core.ui.views.preferences.GroupedItemPosition
@@ -341,7 +340,6 @@ private fun FormattingAction(
         style = GeneralButtonStyle.Text,
         icon = ToolkitIcon.Vector(imageVector = icon),
         contentDescription = contentDescription,
-        measurements = ButtonMeasurements.ExtraSmall,
         iconSize = SizeConstants.ButtonIconSize,
     )
 }